LPS-197867 Inline DDMExpressionFunctionMetadataHelper methods into DDMFormBuilderSettingsRetrieverHelperImpl, fix DDMExpressionFunctionMetadataHelperTest test and rename it to DDMFormBuilderSettingsRetrieverHelperImplTest, since the helper class does no longer exist
