+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/java/com/liferay/dynamic/data/mapping/form/builder/internal/settings/DDMFormBuilderSettingsRetrieverHelperImpl.java
@@ -6,7 +6,9 @@
 package com.liferay.dynamic.data.mapping.form.builder.internal.settings;
 
 import com.liferay.dynamic.data.mapping.constants.DDMStructureConstants;
-import com.liferay.dynamic.data.mapping.form.builder.internal.helper.DDMExpressionFunctionMetadataHelper;
+import com.liferay.dynamic.data.mapping.expression.DDMExpressionFunction;
+import com.liferay.dynamic.data.mapping.expression.DDMExpressionFunctionRegistry;
+import com.liferay.dynamic.data.mapping.form.builder.internal.util.DDMExpressionFunctionMetadata;
 import com.liferay.dynamic.data.mapping.model.DDMForm;
 import com.liferay.dynamic.data.mapping.model.DDMStructure;
 import com.liferay.dynamic.data.mapping.service.DDMStructureService;
@@ -19,11 +21,24 @@ import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.json.JSONSerializer;
+import com.liferay.portal.kernel.language.Language;
+import com.liferay.portal.kernel.util.AggregateResourceBundle;
+import com.liferay.portal.kernel.util.LinkedHashMapBuilder;
 import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.kernel.util.ResourceBundleUtil;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 
+import java.lang.reflect.Method;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
+import java.util.Objects;
+import java.util.ResourceBundle;
 
 import javax.servlet.Servlet;
 import javax.servlet.ServletConfig;
@@ -143,8 +158,7 @@ public class DDMFormBuilderSettingsRetrieverHelperImpl
 		JSONSerializer jsonSerializer = _jsonFactory.createJSONSerializer();
 
 		return jsonSerializer.serializeDeep(
-			_ddmExpressionFunctionMetadataHelper.
-				getDDMExpressionFunctionsMetadata(locale));
+			_getDDMExpressionFunctionsMetadata(locale));
 	}
 
 	@Override
@@ -153,6 +167,186 @@ public class DDMFormBuilderSettingsRetrieverHelperImpl
 
 		return jsonSerializer.serializeDeep(
 			_spiDDMFormRuleConverter.convert(ddmForm.getDDMFormRules()));
+	}
+
+	protected void populateCustomDDMExpressionFunctionsMetadata(
+		Map<String, List<DDMExpressionFunctionMetadata>>
+			ddmExpressionFunctionMetadatasMap,
+		Locale locale) {
+
+		Map<String, DDMExpressionFunction> customDDMExpressionFunctions =
+			_ddmExpressionFunctionRegistry.getCustomDDMExpressionFunctions();
+
+		for (Map.Entry<String, DDMExpressionFunction> entry :
+				customDDMExpressionFunctions.entrySet()) {
+
+			Method method = null;
+
+			DDMExpressionFunction ddmExpressionFunction = entry.getValue();
+
+			Class<?> clazz = ddmExpressionFunction.getClass();
+
+			for (Method curMethod : clazz.getMethods()) {
+				if (Objects.equals(curMethod.getName(), "apply") &&
+					Objects.equals(curMethod.getReturnType(), Boolean.class)) {
+
+					method = curMethod;
+
+					break;
+				}
+			}
+
+			if (method == null) {
+				continue;
+			}
+
+			int parameterCount = method.getParameterCount();
+
+			if (parameterCount > 2) {
+				continue;
+			}
+
+			String label = ddmExpressionFunction.getLabel(locale);
+
+			if (Validator.isNull(label)) {
+				label = entry.getKey();
+			}
+
+			_addDDMExpressionFunctionMetadata(
+				ddmExpressionFunctionMetadatasMap,
+				new DDMExpressionFunctionMetadata(
+					entry.getKey(), label, _TYPE_BOOLEAN,
+					_getParameterClassNames(parameterCount, _TYPE_NUMBER)));
+			_addDDMExpressionFunctionMetadata(
+				ddmExpressionFunctionMetadatasMap,
+				new DDMExpressionFunctionMetadata(
+					entry.getKey(), label, _TYPE_BOOLEAN,
+					_getParameterClassNames(parameterCount, _TYPE_TEXT)));
+		}
+	}
+
+	protected void populateDDMExpressionFunctionsMetadata(
+		Map<String, List<DDMExpressionFunctionMetadata>>
+			ddmExpressionFunctionMetadatasMap,
+		ResourceBundle resourceBundle) {
+
+		_addDDMExpressionFunctionMetadata(
+			ddmExpressionFunctionMetadatasMap,
+			new DDMExpressionFunctionMetadata(
+				"belongs-to", _language.get(resourceBundle, "belongs-to"),
+				_TYPE_BOOLEAN, new String[] {_TYPE_USER, _TYPE_LIST}));
+		_addDDMExpressionFunctionMetadata(
+			ddmExpressionFunctionMetadatasMap,
+			new DDMExpressionFunctionMetadata(
+				"equals-to", _language.get(resourceBundle, "is-equal-to"),
+				_TYPE_BOOLEAN, new String[] {_TYPE_BOOLEAN, _TYPE_BOOLEAN}));
+
+		for (Map.Entry<String, String> entry : _binaryFunctions.entrySet()) {
+			_addDDMExpressionFunctionMetadata(
+				ddmExpressionFunctionMetadatasMap,
+				new DDMExpressionFunctionMetadata(
+					entry.getKey(),
+					_language.get(resourceBundle, entry.getValue()),
+					_TYPE_BOOLEAN, new String[] {_TYPE_NUMBER, _TYPE_NUMBER}));
+			_addDDMExpressionFunctionMetadata(
+				ddmExpressionFunctionMetadatasMap,
+				new DDMExpressionFunctionMetadata(
+					entry.getKey(),
+					_language.get(resourceBundle, entry.getValue()),
+					_TYPE_BOOLEAN, new String[] {_TYPE_TEXT, _TYPE_TEXT}));
+		}
+
+		for (Map.Entry<String, String> entry :
+				_numberBinaryFunctions.entrySet()) {
+
+			_addDDMExpressionFunctionMetadata(
+				ddmExpressionFunctionMetadatasMap,
+				new DDMExpressionFunctionMetadata(
+					entry.getKey(),
+					_language.get(resourceBundle, entry.getValue()),
+					_TYPE_BOOLEAN, new String[] {_TYPE_NUMBER, _TYPE_NUMBER}));
+		}
+
+		for (Map.Entry<String, String> entry :
+				_textBinaryFunctions.entrySet()) {
+
+			_addDDMExpressionFunctionMetadata(
+				ddmExpressionFunctionMetadatasMap,
+				new DDMExpressionFunctionMetadata(
+					entry.getKey(),
+					_language.get(resourceBundle, entry.getValue()),
+					_TYPE_BOOLEAN, new String[] {_TYPE_TEXT, _TYPE_TEXT}));
+		}
+
+		for (Map.Entry<String, String> entry : _unaryFunctions.entrySet()) {
+			_addDDMExpressionFunctionMetadata(
+				ddmExpressionFunctionMetadatasMap,
+				new DDMExpressionFunctionMetadata(
+					entry.getKey(),
+					_language.get(resourceBundle, entry.getValue()),
+					_TYPE_BOOLEAN, new String[] {_TYPE_NUMBER}));
+			_addDDMExpressionFunctionMetadata(
+				ddmExpressionFunctionMetadatasMap,
+				new DDMExpressionFunctionMetadata(
+					entry.getKey(),
+					_language.get(resourceBundle, entry.getValue()),
+					_TYPE_BOOLEAN, new String[] {_TYPE_TEXT}));
+		}
+	}
+
+	private void _addDDMExpressionFunctionMetadata(
+		Map<String, List<DDMExpressionFunctionMetadata>>
+			ddmExpressionFunctionMetadatasMap,
+		DDMExpressionFunctionMetadata ddmExpressionFunctionMetadata) {
+
+		String firstParameterClassName =
+			ddmExpressionFunctionMetadata.getParameterClassNames()[0];
+
+		List<DDMExpressionFunctionMetadata> ddmExpressionFunctionMetadatas =
+			ddmExpressionFunctionMetadatasMap.get(firstParameterClassName);
+
+		if (ddmExpressionFunctionMetadatas == null) {
+			ddmExpressionFunctionMetadatas = new ArrayList<>();
+
+			ddmExpressionFunctionMetadatasMap.put(
+				firstParameterClassName, ddmExpressionFunctionMetadatas);
+		}
+
+		ddmExpressionFunctionMetadatas.add(ddmExpressionFunctionMetadata);
+	}
+
+	private Map<String, List<DDMExpressionFunctionMetadata>>
+		_getDDMExpressionFunctionsMetadata(Locale locale) {
+
+		Map<String, List<DDMExpressionFunctionMetadata>>
+			ddmExpressionFunctionMetadatasMap = new HashMap<>();
+
+		populateCustomDDMExpressionFunctionsMetadata(
+			ddmExpressionFunctionMetadatasMap, locale);
+		populateDDMExpressionFunctionsMetadata(
+			ddmExpressionFunctionMetadatasMap, _getResourceBundle(locale));
+
+		return ddmExpressionFunctionMetadatasMap;
+	}
+
+	private String[] _getParameterClassNames(
+		int parameterCount, String parameterClassName) {
+
+		String[] parameterClassNames = new String[parameterCount];
+
+		Arrays.fill(parameterClassNames, parameterClassName);
+
+		return parameterClassNames;
+	}
+
+	private ResourceBundle _getResourceBundle(Locale locale) {
+		ResourceBundle portalResourceBundle = _portal.getResourceBundle(locale);
+
+		ResourceBundle portletResourceBundle = ResourceBundleUtil.getBundle(
+			"content.Language", locale, getClass());
+
+		return new AggregateResourceBundle(
+			portletResourceBundle, portalResourceBundle);
 	}
 
 	private String _getServletContextPath(Servlet servlet) {
@@ -165,6 +359,45 @@ public class DDMFormBuilderSettingsRetrieverHelperImpl
 		return proxyPath.concat(servletContext.getContextPath());
 	}
 
+	private static final String _TYPE_BOOLEAN = "boolean";
+
+	private static final String _TYPE_LIST = "list";
+
+	private static final String _TYPE_NUMBER = "number";
+
+	private static final String _TYPE_TEXT = "text";
+
+	private static final String _TYPE_USER = "user";
+
+	private static final Map<String, String> _binaryFunctions =
+		LinkedHashMapBuilder.put(
+			"equals-to", "is-equal-to"
+		).put(
+			"not-equals-to", "is-not-equal-to"
+		).build();
+	private static final Map<String, String> _numberBinaryFunctions =
+		LinkedHashMapBuilder.put(
+			"greater-than", "is-greater-than"
+		).put(
+			"greater-than-equals", "is-greater-than-or-equal-to"
+		).put(
+			"less-than", "is-less-than"
+		).put(
+			"less-than-equals", "is-less-than-or-equal-to"
+		).build();
+	private static final Map<String, String> _textBinaryFunctions =
+		LinkedHashMapBuilder.put(
+			"contains", "contains"
+		).put(
+			"not-contains", "does-not-contain"
+		).build();
+	private static final Map<String, String> _unaryFunctions =
+		LinkedHashMapBuilder.put(
+			"is-empty", "is-empty"
+		).put(
+			"not-is-empty", "is-not-empty"
+		).build();
+
 	@Reference(
 		target = "(osgi.http.whiteboard.servlet.name=com.liferay.dynamic.data.mapping.form.builder.internal.servlet.DDMDataProviderInstanceParameterSettingsServlet)"
 	)
@@ -176,8 +409,7 @@ public class DDMFormBuilderSettingsRetrieverHelperImpl
 	private Servlet _ddmDataProviderInstancesServlet;
 
 	@Reference
-	private DDMExpressionFunctionMetadataHelper
-		_ddmExpressionFunctionMetadataHelper;
+	private DDMExpressionFunctionRegistry _ddmExpressionFunctionRegistry;
 
 	@Reference(
 		target = "(osgi.http.whiteboard.servlet.name=com.liferay.dynamic.data.mapping.form.builder.internal.servlet.DDMFieldSetDefinitionServlet)"
@@ -204,6 +436,9 @@ public class DDMFormBuilderSettingsRetrieverHelperImpl
 
 	@Reference
 	private JSONFactory _jsonFactory;
+
+	@Reference
+	private Language _language;
 
 	@Reference
 	private Portal _portal;
